refactor: comments controller
